DdmlibAdbServer wait for device list initialized
In com.android.ddmlib.DeviceMonitor, its `mDeviceListMonitorTask` runs
in another thread, to update `mDevices` without blocking original
thread.

In bundletool, we get devices as soon as adb connected. It is possible
that DeviceMonitor has not detect anything before we call `getDevices`.

To deal with that, we block thread for a bit.

to resolve #11